+++ src/js/appContainer.jsx
@@ -1,13 +1,25 @@
-import React, { Fragment } from "react";
+import React, { useEffect, Fragment } from "react";
+import { connect } from "react-redux";
 import classNames from "classnames";
+import PropType from "prop-types";
 import { makeStyles } from "@material-ui/core";
 import AppFooter from "./components/appFooter";
 import LeftNavBar from "./components/leftNavBar";
 import NavTabs from "./components/sections/tabs";
+import { fetchArticles } from "./components/lists/articles/state/actions";
 import { appContainerStyles } from "./styles";
 
+const propTypes = { callFetchArticles: PropType.func };
 const useStyles = makeStyles(() => appContainerStyles);
-const AppContainer = () => {
+const AppContainer = ({ callFetchArticles }) => {
+  useEffect(() => {
+    console.info("useEffect: ", typeof callFetchArticles);
+    if (typeof callFetchArticles === "function") {
+      callFetchArticles();
+      console.info("hit");
+    }
+  }, [callFetchArticles]);
+
   const classes = useStyles();
   return (
     <Fragment>
@@ -29,4 +41,7 @@ const AppContainer = () => {
   );
 };
 
-export default AppContainer;
+AppContainer.propTypes = propTypes;
+const mapStateToProps = () => ({});
+const mapDispatchToProps = { callFetchArticles: fetchArticles };
+export default connect(mapStateToProps, mapDispatchToProps)(AppContainer);